// src/app/Pages/Master Module/Models/pincode.ts
export class pincode {
    id: number;
    district_id: any;
    state_id: any;
    country_id: any;
    area_name: string;
    pincode_number: any;
    seq_no: number = 0;
    is_active: any = 1;
    client_id: any;

    district_name: any;
    state_name: any;
    country_name: any;

    division_name: string;
    circle_name: string;
    taluka: string;
    sub_office: string;
    head_office: string;
    longitude: any;
    latitude: any;

}